+++ src/rendering/light.rs
@@ -1,7 +1,7 @@
 
 // Due to uniforms requiring 16 byte (4 float) spacing, its needed to use pading
 // position: [f32; 3]
-// __padding: [f32; 1]
+// _padding: [f32; 1]
 // --> 16
 // https://sotrh.github.io/learn-wgpu/intermediate/tutorial10-lighting/#the-blinn-phong-model
 // https://www.w3.org/TR/WGSL/#alignment-and-size